Bugfix: Copy instead of symlinking during setup.
The bugfix to Issue #2 introduced its own problem, so now we just copy the binaries instead of messing around with symbolic and hard links. Hopefully this time the bug is squashed for good. This commit also adds a new test trace that was broken by the prior bugfix so the same mistake isn't made again.