Bugfix: Avoid nondeterminism introduced by solver.unique. (#13)
When state.solver.unique() is called, under the hood angr invokes the method eval_upto(2) to see whether more than one satisfiable value exists. This is cached as a set whose ordering in some Python versions depends on a hash seed, causing the subsequent call to eval() to be nondeterministic and yield inconsistent results.

Our solution is to instead use state.solver.is_true(), which returns True if an expression can be determined to be True without needing to invoke the solver. This can be used to determine whether the current symbolic variable is obviously equivalent to a prior instance of itself, namely in the case that the constraints haven't changed, which suits our goals in the fixed plugins. This fix has two advantages: 1) it is more efficient because the solver is no longer being invoked and 2) it more gracefully handles symbolic variables with more than 1 satisfiable value compared to using unique() + eval().