Simulation procedures for wide-characters, libpthreads, and unrar. (#28)
This commit adds 21 new simulation procedures to ARCUS. Specifically, it adds procedures to simulate most of libc's functions for handling wide-characters. It also includes "null" procedures for skipping functions in unrar and libpthreads that are not relevant to ARCUS' current analyses. These significantly reduce analysis time and constraint complexity for relevant target programs.