Add RISC-V workaround for tracking call depth.
As of v9.2.25 of angr, pyvex does not correctly recognize all riscv64 instructions used to make calls, marking some wrongly as jumpkind Ijk_Boring, which can cause ARCUS to end analysis early. As a workaround, we now recognize when the state's architecture is RISC-V and check manually for the jump-and-link instructions to correctly track call depth.