Bugfix: In the loop exploration plugin, do not rely on archinfo.call_pushes_ret or the BP register to determine whether there's a return pointer on stack.
Prior to this commit, when deciding whether to apply a heuristic designed to detect when a return pointer has been overwritten by an out-of-bounds write, the loop exploration plugin would check whether the target architecture's call instruction automatically pushes return pointers. This is not a reliable way to detect whether there are return pointers on the stack. For example, while aarch64's call instruction places the return address into a register, most programs immediately store it to the stack.

Additionally, the plugin was relying on the BP register in some cases to locate the current stack frame, but this is not reliable either because some programs don't use the BP register, which can cause the plugin to exit loops too soon.